+++ src/problem_types/cutting_stock/cutting_stock.jl
@@ -9,3 +9,4 @@
 include("standard.jl")
 include("setup_cost.jl")
 include("due_dates.jl")
+include("integer_patterns.jl")